Extract performance logger
