+++ hahomematic/decorators.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import asyncio
 from collections.abc import Awaitable, Callable
 from contextvars import Token
 from datetime import datetime
@@ -58,25 +59,32 @@ def service(
                 return cast(T, no_raise_return)
             finally:
                 if do_measure_performance:
-                    delta = (datetime.now() - start).total_seconds()
-                    caller = str(args[0]) if len(args) > 0 else ""
-
-                    iface: str = ""
-                    if interface := str(kwargs.get("interface", "")):
-                        iface = f"interface: {interface}"
-                    if interface_id := kwargs.get("interface_id", ""):
-                        iface = f"interface_id: {interface_id}"
-
-                    message = f"Execution of {func.__name__} took {delta}s from {caller}"
-                    if iface:
-                        message += f"/{iface}"
-
-                    _LOGGER.info(message)
+                    _log_performance_message(func, start, *args, **kwargs)
 
         setattr(service_wrapper, "ha_service", True)
         return service_wrapper
 
     return service_decorator
+
+
+def _log_performance_message(
+    func: Callable, start: datetime, *args: P.args, **kwargs: P.kwargs
+) -> None:
+    """Log the performance message."""
+    delta = (datetime.now() - start).total_seconds()
+    caller = str(args[0]) if len(args) > 0 else ""
+
+    iface: str = ""
+    if interface := str(kwargs.get("interface", "")):
+        iface = f"interface: {interface}"
+    if interface_id := kwargs.get("interface_id", ""):
+        iface = f"interface_id: {interface_id}"
+
+    message = f"Execution of {func.__name__} took {delta}s from {caller}"
+    if iface:
+        message += f"/{iface}"
+
+    _LOGGER.info(message)
 
 
 def get_service_calls(obj: object) -> dict[str, Callable]:
@@ -88,3 +96,35 @@ def get_service_calls(obj: object) -> dict[str, Callable]:
         and callable(getattr(obj, name))
         and hasattr(getattr(obj, name), "ha_service")
     }
+
+
+def measure_execution_time[_CallableT: Callable[..., Any]](func: _CallableT) -> _CallableT:
+    """Decorate function to measure the function execution time."""
+
+    is_enabled = _LOGGER.isEnabledFor(level=logging.DEBUG)
+
+    @wraps(func)
+    async def async_measure_wrapper(*args: Any, **kwargs: Any) -> Any:
+        """Wrap method."""
+        if is_enabled:
+            start = datetime.now()
+        try:
+            return await func(*args, **kwargs)
+        finally:
+            if is_enabled:
+                _log_performance_message(func, start, *args, **kwargs)
+
+    @wraps(func)
+    def measure_wrapper(*args: Any, **kwargs: Any) -> Any:
+        """Wrap method."""
+        if is_enabled:
+            start = datetime.now()
+        try:
+            return func(*args, **kwargs)
+        finally:
+            if is_enabled:
+                _log_performance_message(func, start, *args, **kwargs)
+
+    if asyncio.iscoroutinefunction(func):
+        return async_measure_wrapper  # type: ignore[return-value]
+    return measure_wrapper  # type: ignore[return-value]
